Big count support for datatypes
Utilize the count and disp arrays for type-punning of int and MPI_Count
arguments to datatype creation functions. Pack integer or MPI_Count
depending on what is needed when packing a datatype. Adjust places
where bigcount support for datatpyes was missing.

+++ ompi/datatype/ompi_datatype_create.c
@@ -71,7 +71,7 @@ static void __ompi_datatype_release(ompi_datatype_t * datatype)
 
 OBJ_CLASS_INSTANCE(ompi_datatype_t, opal_datatype_t, __ompi_datatype_allocate, __ompi_datatype_release);
 
-ompi_datatype_t * ompi_datatype_create( int32_t expectedSize )
+ompi_datatype_t * ompi_datatype_create( size_t expectedSize )
 {
     int ret;
     ompi_datatype_t * datatype = (ompi_datatype_t*)OBJ_NEW(ompi_datatype_t);
